refactor(test): Split monolithic SmsParserTest into logical files
