+++ app/src/test/java/io/pm/finlight/CustomRuleParserTest.kt
@@ -1,0 +1,11 @@
+// =================================================================================
+// FILE: ./app/src/test/java/io/pm/finlight/CustomRuleParserTest.kt
+// REASON: NEW FILE - A placeholder test file for validating user-defined custom
+// parsing rules. This separates the testing of core parser logic from the
+// user-configurable logic.
+// =================================================================================
+package io.pm.finlight
+
+// TODO: Add tests for custom, user-defined parsing rules here.
+class CustomRuleParserTest : BaseSmsParserTest() {
+}
